refactor: move revoke types to seperate type file
