+++ src/navigation/types.ts
@@ -43,7 +43,7 @@ import { type ScrollView } from 'react-native';
 import { type HlTrade, type PerpMarket, type PerpsPosition, type TriggerOrderSource, type TriggerOrderType } from '@/features/perps/types';
 import { type PolymarketPosition } from '@/features/polymarket/types';
 import { type PolymarketEvent, type PolymarketMarket, type PolymarketMarketEvent } from '@/features/polymarket/types/polymarket-event';
-import { type RevokeReason } from '@/screens/delegation/RevokeDelegationPanel';
+import { type RevokeReason } from '@/screens/delegation/types';
 
 export type PortalSheetProps = {
   children: React.FC;
